+++ application/src/main/java/br/com/jpcchaves/application/resources/CreateResource.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.infrastructure.controller.resources.base;
+package br.com.jpcchaves.application.resources;
 
 public interface CreateResource<Req, Res> {
   Res create(Req req);
