+++ application/src/main/java/br/com/jpcchaves/application/resources/DeleteResource.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.infrastructure.controller.resources.base;
+package br.com.jpcchaves.application.resources;
 
 public interface DeleteResource<PK, Res> {
   Res deleteById(PK id);
