+++ application/src/main/java/br/com/jpcchaves/application/resources/GetByIdResource.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.infrastructure.controller.resources.base;
+package br.com.jpcchaves.application.resources;
 
 public interface GetByIdResource<PK, Res> {
   Res getById(PK id);
