+++ application/src/main/java/br/com/jpcchaves/application/resources/ListResource.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.infrastructure.controller.resources.base;
+package br.com.jpcchaves.application.resources;
 
 import java.util.List;
