+++ application/src/main/java/br/com/jpcchaves/application/resources/UpdateResource.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.infrastructure.controller.resources.base;
+package br.com.jpcchaves.application.resources;
 
 public interface UpdateResource<PK, Req, Res> {
   Res update(PK id, Req req);
